Traditional Chinese Medicine expert/requirements.txt
flask
flask-cors
openai
langchain-community
langchain-huggingface
chromadb
torch
torchvision
pillow
